Set a full target triple variable
The arch triple does not include the deployment target on
Apple-platforms. Passing it directly to the Swift compiler results in an
error because the deployment target is too low for Swift by default.
For the most part, it's appending the deployment target version on the
target triple except for Catalyst and simulator builds, where the
deployment version has to go before the `-macabi` and `-simulator` parts
of the triple.