CMake: SwiftCore: Directory Swift Modules
The catalyst support in the driver requires that the swiftmodules for
the catalyst slice live in the swiftmodule directory. CMake currently
doesn't do this natively so this patch works around it by specifying the
module path explicitly. The use of an empty custom command is used to
hook up the dependency graph on the specific target swiftmodule files
correctly.

This is only needed for catalyst builds that this time so limit the
affect of this change to instances where we are building zippered
catalyst binaries.

The nested structure also greatly simplifies the installation story. We
simply need to copy the entire swiftmodule directory into the install
location.