Update-Checkout Tests: Disable commit singing
Disabling commit signing on the local mock repo so that signing tools
don't cause the tests to fail when the signing identity doesn't match
the identity of the committer.

Also marking the config changes as local to ensure that they don't
affect the user's global config accidentally. Technically this isn't
required since local is the default, but this ensures that it is clear
and that we aren't making assumptions about the user's global git
config.