Sema: Fix case where witness thrown error type is a subtype of a type parameter
In b30006837e31689c4eb96d27a00c583024b1b0ea, I changed the `if`
condition here to check for the absence of type variables as well
as type parameters. This is incorrect; the type variables come up
in ValueWitnessRequest, and the type parameters come up in
associated type inference. We want the matching to be more lax
in the former case.

Fixes rdar://149438520.