UsePrespecialized: Don't build a generic mangling for symbols we don't have a pre-specialiation for
This is a compiler performance optimization to not do useless work.

The deleted test case: test/SILOptimizer/prespecialization_with_definition.sil
was added to test an assert that had to be moved and not really the
functionality requirement.

If the:

```
    // If we already have this specialization, reuse it.
    auto PrevF = M.lookUpFunction(ClonedName);
```

is indeed load/performance bearing we should see this in the Onone
benchmarks I am going to run.

NFC intended.