SIL: Use SubstitutionMap::mapIntoTypeExpansionContext() in SILTypeSubstituter
I made a mistake in 47156e006bf0029c59edfd072392ae17045db73d. There was a
call to call to forAbstract() in SILTypeSubstituter that passed in the
wrong subject ype.

This call was inside SILTypeSubstituter's own implementation of replacing
opaque types with underlying types in a substitution map. This duplicates
an existing utility method in SubstitutionMap anyway, so let's just use
that instead.

Fixes rdar://149353285.