Bootstrap: Install dispatch in stage0
The concurrency runtimes require that dispatch is installed or it will
fail to load at launch. The C library is built as part of the
concurrency build but is not installed. We need to install a copy of
dispatch so that programs that link concurrency run.

(cherry picked from commit f36391d4c803f00fc5e4cebca910114dfadbfdd7)