Don't depend on implicit lifetime of Escapable inout
Escapable inout parameters don't have an implicit lifetime that can be
borrowed or copied, because the caller doesn't know what the lifetime of
the new value is, if the callee writes to the parameter. Depending on
the lifetime of an Escapable inout parameter without explicit lifetime
annotation results in an error, so _SwiftifyImport should not do that.