Add `typing` to `update_checkout.py`
As Python 3.5 added support for type hints and we require Python 3.6 or later, let's try to add a type hint for a function so that development experience is improved. Editors with support for Python LSP servers can pick it and stop flagging this function or its invocations as type errors.