[test] IRGen: Adjust FileCheck pattern for `nocapture` refactoring in upstream LLVM
The current pattern does not match a ')' inside the argument list. This
will no longer work because `nocapture` became `captures(none) in
29441e4f5fa5f5c7709f7cf180815ba97f611297 (llvm-project).

Use a wildcard regex instead. Doing so does not compromise the test
because the `unnamed_addr` attribute we are guarding against can appear
only after the argument list.

This is intended to reduce future conflicts with rebranch.