[test] IRGen: Adjust FileCheck patterns for new `captures` attribute in upstream LLVM
The `captures(ret: address, provenance)` parameter attribute is expected
because these functions either immediately return the pointer argument
or pass it as an argument to another such parameter. See
https://llvm.org/docs/LangRef.html#parameter-attributes.

Match it using a wildcard regex, since it is not relevant to these
tests.

This is intended to reduce future conflicts with rebranch.