[test] IRGen: Adjust FileCheck pattern for new `range` attribute in upstream LLVM
The attribute is `range(i8 -127, -128)`. The documentation specifies
that the range can wrap, meaning that for i8 [-127, -128) is equivalent
to [-127, 127). This is expected for a function that simply increments a
i8 value. See https://llvm.org/docs/LangRef.html#parameter-attributes.

Match it using a wildcard regex, since it is not relevant to this test.

This is intended to reduce future conflicts with rebranch.