[Concurrency] Bring back ability to get a queue custom executor in executorForEnqueuedJob
This was a regression and now even if we had a queue set in
job->SchedulerPrivate[Job::DispatchQueueIndex].

Arguably the way we're using that storage needs to be fixed somehow,
since if we're opening this storage to other executors then it means
we could have OTHER things in there, not just a dispatch queueue,
and therefore this cast would be bad.

For now though we can bring back the existing behavior while we work out
with Alastair what the right way to handle this will be -- some marker
bit somewhere that indeed this is a queue stored here?

resolves rdar://150213107